Get WorkflowTaskFailedEvent from full history
